Create CrossingBridge_BFS.py
I am modeling the 'Crossing the Bridge' puzzle which consists of: a bridge will collapse in 17 minutes. 4 people want to cross it before it will collapse. It is a dark night and there is only one torch between them. Only two people can cross at a time. “A” takes a minute to cross. “B” takes 2 minutes, “C” takes 5 minutes and “D” takes 10 minutes. So, How do they all cross before the bridge cross? With the BFS (Breadth First Search) technique, which is one of the blind search techniques.